fix executor_index error.
